🎨 move database connector to manager
