+++ src/commands/config/modules/embeds/index.ts
@@ -10,7 +10,7 @@ import {
 import logger from "../../../../logger";
 
 // Models
-import guildSchema from "../../../../database/schemas/guild";
+import guildSchema from "../../../../models/guild";
 import { SlashCommandSubcommandBuilder } from "@discordjs/builders";
 import getEmbedConfig from "../../../../helpers/getEmbedConfig";
 import getValues from "./components/getValues";
