+++ src/events/guildMemberAdd/audits.ts
@@ -1,7 +1,7 @@
 import logger from "../../logger";
 import { GuildMember, MessageEmbed, TextChannel } from "discord.js";
 
-import guildSchema from "../../database/schemas/guild";
+import guildSchema from "../../models/guild";
 
 import getEmbedConfig from "../../helpers/getEmbedConfig";
