+++ src/events/messageUpdate/audits.ts
@@ -2,7 +2,7 @@
 import logger from "../../logger";
 import { Message, MessageEmbed, TextChannel } from "discord.js";
 
-import guildSchema from "../../database/schemas/guild";
+import guildSchema from "../../models/guild";
 
 import getEmbedConfig from "../../helpers/getEmbedConfig";
