+++ src/index.ts
@@ -4,7 +4,7 @@ import { token, intents } from "./config/discord";
 
 import { Client } from "discord.js"; // discord.js
 
-import database from "./database";
+import * as databaseManager from "./managers/database";
 import * as scheduleManager from "./managers/schedule";
 import * as eventManager from "./managers/event";
 import * as commandManager from "./managers/command";
@@ -16,16 +16,10 @@ const main = async () => {
     intents,
   });
 
-  // Start database manager
-  await database();
-
-  // Start schedule manager
+  // Start managers
+  await databaseManager.start();
   await scheduleManager.start(client);
-
-  // Start command handler
   await commandManager.register(client);
-
-  // Start event handler
   await eventManager.register(client);
 
   // Authorize with Discord's API
