+++ src/managers/database/index.ts
@@ -2,12 +2,12 @@
 import mongoose from "mongoose";
 
 // Dependencies
-import logger from "../logger";
+import logger from "../../logger";
 
 // Configuration
-import { url } from "../config/database";
+import { url } from "../../config/database";
 
-export default async () => {
+export const start = async () => {
   await mongoose.connect(url).then(async (connection) => {
     logger.info(`Connected to database: ${connection.connection.name}`);
   });
