+++ src/models/guild.ts
